Define conf and peer types
Reference https://github.com/pirate/wireguard-docs
